Module oak_services::proto::oak::crypto

Modules

keyset_generate_request
Nested message and enum types in KeysetGenerateRequest.

Structs

AeadDecryptRequest

AeadDecryptResponse

AeadEncryptRequest

AeadEncryptResponse

ComputeMacRequest

ComputeMacResponse

ComputePrfRequest

ComputePrfResponse

DeterministicAeadDecryptRequest

DeterministicAeadDecryptResponse

DeterministicAeadEncryptRequest

DeterministicAeadEncryptResponse

KeyTemplate
Local copy of Tink key template definition, to prevent needing a full dependency on all of Tink's protos.

KeysetBindRequest

KeysetBindResponse

KeysetGenerateRequest

KeysetPublicRequest

KeysetResponse
Common response type for acquiring keyset handles

KeysetUnbindRequest

KmsProxyRequest

SignatureSignRequest

SignatureSignResponse

SignatureVerifyRequest

SignatureVerifyResponse

VerifyMacRequest

VerifyMacResponse

Enums

KeysetFormat

OutputPrefixType