Module oak_services::proto::oak::log

Structs

LogInit

A convenience message that may be used as init message by nodes that only expect to receive a single handle to a log node at startup.

LogMessage

This message defines the data that is passed as a log message to a logging pseudo-node. It provides a “schema” to keep the logging pseudo-node in sync with the internal representation of log messages in Oak SDK.

Enums

Level

Logging levels as defined in https://docs.rs/log/0.4.10/log/enum.Level.html. UNKNOWN_LEVEL is added as the default value.